+++ Scripts/Runtime/Collections/DeferredNativeArray.cs
@@ -8,35 +8,16 @@ using Unity.Jobs;
 
 namespace Anvil.Unity.DOTS.Collections
 {
-    [BurstCompatible]
-    [StructLayout(LayoutKind.Sequential)]
-    internal unsafe struct DeferredNativeArrayBufferInfo
-    {
-        public static readonly int SIZE = UnsafeUtility.SizeOf<DeferredNativeArrayBufferInfo>();
-        public static readonly int ALIGNMENT = UnsafeUtility.AlignOf<DeferredNativeArrayBufferInfo>();
-
-        [NativeDisableUnsafePtrRestriction] public void* Buffer;
-        public int Length;
-        public int MaxIndex;
-        public DeferredNativeArrayState State;
-    }
-
-    internal enum DeferredNativeArrayState
-    {
-        Placeholder,
-        Created
-    }
-
     /// <summary>
     /// A native collection similar to <see cref="NativeArray{T}"/> but intended for use in a deferred context.
     /// Useful for cases where a job that hasn't finished yet will determine the length of the array.
     ///
-    /// Initially upon creation, this collection is an empty array that cannot be interacted with.
+    /// Upon creation, this collection is an empty array that cannot be interacted with.
     /// You can pass it into a job as a <see cref="NativeArray{T}"/> using <see cref="AsDeferredJobArray"/>.
     /// That array will be empty and only populated later on when the job is run.
     ///
-    /// To populate you pass this <see cref="DeferredNativeArray{T}"/> into a job and then call
-    /// <see cref="DeferredCreate"/> when you know what the length is which will give you the right
+    /// To populate, pass this <see cref="DeferredNativeArray{T}"/> into a job and then call
+    /// <see cref="DeferredCreate"/> when the length of the array is known. This will allocate the right
     /// size of <see cref="NativeArray{T}"/> to act on in your job and populate.
     ///
     /// In your later jobs that used <see cref="AsDeferredJobArray"/> it will be populated properly.
@@ -53,8 +34,33 @@ namespace Anvil.Unity.DOTS.Collections
     public struct DeferredNativeArray<T> : INativeDisposable
         where T : struct
     {
+        //*************************************************************************************************************
+        // INTERNAL STRUCTS
+        //*************************************************************************************************************
+        
+        [BurstCompatible]
+        [StructLayout(LayoutKind.Sequential)]
+        internal unsafe struct BufferInfo
+        {
+            // ReSharper disable once MemberHidesStaticFromOuterClass
+            public static readonly int SIZE = UnsafeUtility.SizeOf<BufferInfo>();
+
+            // ReSharper disable once MemberHidesStaticFromOuterClass
+            public static readonly int ALIGNMENT = UnsafeUtility.AlignOf<BufferInfo>();
+
+            [NativeDisableUnsafePtrRestriction] public void* Buffer;
+            public int Length;
+        }
+        
+        //*************************************************************************************************************
+        // STATIC HELPERS
+        //*************************************************************************************************************
+
         private static readonly int SIZE = UnsafeUtility.SizeOf<T>();
         private static readonly int ALIGNMENT = UnsafeUtility.AlignOf<T>();
+
+        // ReSharper disable once StaticMemberInGenericType
+        private static int s_StaticSafetyId;
 
         [BurstDiscard]
         private static void InitStaticSafetyId(ref AtomicSafetyHandle handle)
@@ -68,7 +74,7 @@ namespace Anvil.Unity.DOTS.Collections
         }
 
         [BurstDiscard]
-        private static void IsUnmanagedAndThrow()
+        private static void AssertValidElementType()
         {
             if (UnsafeUtility.IsValidNativeContainerElementType<T>())
             {
@@ -85,27 +91,25 @@ namespace Anvil.Unity.DOTS.Collections
                 throw new ArgumentException("Allocator must be Temp, TempJob or Persistent", nameof(allocator));
             }
 
-            IsUnmanagedAndThrow();
+            AssertValidElementType();
             array = new DeferredNativeArray<T>();
-            void* initialBuffer = UnsafeUtility.Malloc(SIZE, ALIGNMENT, allocator);
-            array.m_BufferInfo = (DeferredNativeArrayBufferInfo*)UnsafeUtility.Malloc(DeferredNativeArrayBufferInfo.SIZE,
-                                                                                      DeferredNativeArrayBufferInfo.ALIGNMENT,
-                                                                                      allocator);
+            array.m_BufferInfo = (BufferInfo*)UnsafeUtility.Malloc(BufferInfo.SIZE,
+                                                                   BufferInfo.ALIGNMENT,
+                                                                   allocator);
             array.m_BufferInfo->Length = 0;
-            array.m_BufferInfo->MaxIndex = 0;
-            array.m_BufferInfo->Buffer = initialBuffer;
-            array.m_BufferInfo->State = DeferredNativeArrayState.Placeholder;
+            array.m_BufferInfo->Buffer = null;
 
             array.m_Allocator = allocator;
 
             DisposeSentinel.Create(out array.m_Safety, out array.m_DisposeSentinel, 1, allocator);
             InitStaticSafetyId(ref array.m_Safety);
         }
+        
+        //*************************************************************************************************************
+        // NATIVE COLLECTION
+        //*************************************************************************************************************
 
-        // ReSharper disable once StaticMemberInGenericType
-        private static int s_StaticSafetyId;
-
-        [NativeDisableUnsafePtrRestriction] internal unsafe DeferredNativeArrayBufferInfo* m_BufferInfo;
+        [NativeDisableUnsafePtrRestriction] internal unsafe BufferInfo* m_BufferInfo;
 
         [NativeSetClassTypeToNullOnSchedule] private DisposeSentinel m_DisposeSentinel;
 
@@ -117,7 +121,17 @@ namespace Anvil.Unity.DOTS.Collections
         /// </summary>
         public unsafe bool IsCreated
         {
-            get => m_BufferInfo != null && m_BufferInfo->Buffer != null;
+            get => m_BufferInfo != null;
+        }
+        
+        /// <summary>
+        /// If <see cref="DeferredCreate"/> has NOT yet been called. (Length is 0)
+        /// This will return true.
+        /// False if the array was properly sized and Length is greater than 0.
+        /// </summary>
+        public unsafe bool IsPendingDeferredCreate
+        {
+            get => m_BufferInfo->Length == 0;
         }
 
         /// <summary>
@@ -135,20 +149,17 @@ namespace Anvil.Unity.DOTS.Collections
         [WriteAccessRequired]
         public unsafe void Dispose()
         {
-            if (!UnsafeUtility.IsValidAllocator(m_Allocator))
-            {
-                throw new InvalidOperationException("The DeferredNativeArray can not be Disposed because it was not allocated with a valid allocator.");
-            }
-
-            DisposeSentinel.Dispose(ref m_Safety, ref m_DisposeSentinel);
-
-            if (m_BufferInfo == null)
+            if (!IsCreated)
             {
                 return;
             }
-
-            UnsafeUtility.Free(m_BufferInfo->Buffer, m_Allocator);
-            m_BufferInfo->Buffer = null;
+            
+            DisposeSentinel.Dispose(ref m_Safety, ref m_DisposeSentinel);
+            if (m_BufferInfo->Buffer != null)
+            {
+                UnsafeUtility.Free(m_BufferInfo->Buffer, m_Allocator);
+                m_BufferInfo->Buffer = null;
+            }
             UnsafeUtility.Free(m_BufferInfo, m_Allocator);
             m_BufferInfo = null;
         }
@@ -163,6 +174,10 @@ namespace Anvil.Unity.DOTS.Collections
         /// <returns>A <see cref="JobHandle"/> for when the disposal is complete.</returns>
         public unsafe JobHandle Dispose(JobHandle inputDeps)
         {
+            if (!IsCreated)
+            {
+                return default;
+            }
             DisposeSentinel.Clear(ref m_DisposeSentinel);
             DisposeJob disposeJob = new DisposeJob(m_BufferInfo, m_Allocator);
             JobHandle jobHandle = disposeJob.Schedule(inputDeps);
@@ -171,21 +186,7 @@ namespace Anvil.Unity.DOTS.Collections
             m_BufferInfo = null;
             return jobHandle;
         }
-
-        [Conditional("ENABLE_UNITY_COLLECTIONS_CHECKS")]
-        [BurstDiscard]
-        private unsafe void AssertForDeferredCreate()
-        {
-            Debug.Assert(m_BufferInfo->State == DeferredNativeArrayState.Placeholder, $"{nameof(DeferredNativeArray<T>)} has already been created! Cannot call {nameof(DeferredCreate)} more than once.");
-        }
-
-        [Conditional("ENABLE_UNITY_COLLECTIONS_CHECKS")]
-        [BurstDiscard]
-        private unsafe void AssertForAsDeferredJobArray()
-        {
-            Debug.Assert(m_BufferInfo->State == DeferredNativeArrayState.Created, $"{nameof(DeferredNativeArray<T>)} has already called {nameof(DeferredCreate)}! Cannot call {nameof(AsDeferredJobArray)} after.");
-        }
-
+        
         /// <summary>
         /// Creates the actual array for when you know the length.
         /// Usually inside a job.
@@ -203,14 +204,10 @@ namespace Anvil.Unity.DOTS.Collections
             {
                 UnsafeUtility.MemClear(newMemory, size);
             }
-
-            //Free the original memory
-            UnsafeUtility.Free(m_BufferInfo->Buffer, m_Allocator);
+            
             //Update the buffer info
             m_BufferInfo->Length = newLength;
-            m_BufferInfo->MaxIndex = newLength - 1;
             m_BufferInfo->Buffer = newMemory;
-            m_BufferInfo->State = DeferredNativeArrayState.Created;
 
             //Return an actual NativeArray so it's familiar to use and we don't have to reimplement the same api and functionality
             NativeArray<T> array = NativeArrayUnsafeUtility.ConvertExistingDataToNativeArray<T>(m_BufferInfo->Buffer, newLength, m_Allocator);
@@ -220,7 +217,14 @@ namespace Anvil.Unity.DOTS.Collections
 #endif
             return array;
         }
-
+        
+        [Conditional("ENABLE_UNITY_COLLECTIONS_CHECKS")]
+        [BurstDiscard]
+        private unsafe void AssertForDeferredCreate()
+        {
+            Debug.Assert(m_BufferInfo->Length == 0, $"{nameof(DeferredNativeArray<T>)} has already been created! Cannot call {nameof(DeferredCreate)} more than once.");
+        }
+        
         /// <summary>
         /// Returns a <see cref="NativeArray{T}"/> for use in a job.
         /// Initially this <see cref="NativeArray{T}"/> will not have anything in it but later on after
@@ -249,6 +253,13 @@ namespace Anvil.Unity.DOTS.Collections
 
             return array;
         }
+        
+        [Conditional("ENABLE_UNITY_COLLECTIONS_CHECKS")]
+        [BurstDiscard]
+        private unsafe void AssertForAsDeferredJobArray()
+        {
+            Debug.Assert(m_BufferInfo->Length > 0, $"You are trying to call {nameof(AsDeferredJobArray)} after {nameof(DeferredCreate)} has already been called which is not allowed.");
+        }
 
         //*************************************************************************************************************
         // JOBS
@@ -257,10 +268,10 @@ namespace Anvil.Unity.DOTS.Collections
         [BurstCompile]
         private readonly unsafe struct DisposeJob : IJob
         {
-            [NativeDisableUnsafePtrRestriction] private readonly DeferredNativeArrayBufferInfo* m_BufferInfo;
+            [NativeDisableUnsafePtrRestriction] private readonly BufferInfo* m_BufferInfo;
             private readonly Allocator m_Allocator;
 
-            public DisposeJob(DeferredNativeArrayBufferInfo* bufferInfo, Allocator allocator)
+            public DisposeJob(BufferInfo* bufferInfo, Allocator allocator)
             {
                 m_BufferInfo = bufferInfo;
                 m_Allocator = allocator;
@@ -270,64 +281,12 @@ namespace Anvil.Unity.DOTS.Collections
             {
                 //This dispose job just handles freeing the memory, the other aspects of the collection were already
                 //when this job was scheduled because it requires main thread access
-                UnsafeUtility.Free(m_BufferInfo->Buffer, m_Allocator);
+                if (m_BufferInfo->Buffer != null)
+                {
+                    UnsafeUtility.Free(m_BufferInfo->Buffer, m_Allocator);
+                }
                 UnsafeUtility.Free(m_BufferInfo, m_Allocator);
             }
-        }
-    }
-
-    [BurstCompatible]
-    public static unsafe class DeferredNativeArrayUnsafeUtility
-    {
-        /// <summary>
-        /// Gets the pointer to <see cref="DeferredNativeArrayBufferInfo"/> struct.
-        /// Will check the safety handle
-        /// </summary>
-        /// <param name="deferredNativeArray">The <see cref="DeferredNativeArray{T}"/> to get the pointer from</param>
-        /// <typeparam name="T">The type of <see cref="DeferredNativeArray{T}"/> it is.</typeparam>
-        /// <returns>The pointer</returns>
-        public static void* GetBufferInfo<T>(ref DeferredNativeArray<T> deferredNativeArray)
-            where T : struct
-        {
-            AtomicSafetyHandle.CheckWriteAndThrow(deferredNativeArray.m_Safety);
-            return deferredNativeArray.m_BufferInfo;
-        }
-
-        /// <summary>
-        /// Gets the pointer to <see cref="DeferredNativeArrayBufferInfo"/> struct.
-        /// Will NOT check the safety handle
-        /// </summary>
-        /// <param name="deferredNativeArray">The <see cref="DeferredNativeArray{T}"/> to get the pointer from</param>
-        /// <typeparam name="T">The type of <see cref="DeferredNativeArray{T}"/> it is.</typeparam>
-        /// <returns>The pointer</returns>
-        public static void* GetBufferInfoUnchecked<T>(ref DeferredNativeArray<T> deferredNativeArray)
-            where T : struct
-        {
-            return deferredNativeArray.m_BufferInfo;
-        }
-
-        /// <summary>
-        /// Gets the safety handle for a <see cref="DeferredNativeArray{T}"/>
-        /// </summary>
-        /// <param name="deferredNativeArray">The instance to get the safety handle from.</param>
-        /// <typeparam name="T">The type of <see cref="DeferredNativeArray{T}"/></typeparam>
-        /// <returns>An <see cref="AtomicSafetyHandle"/> instance</returns>
-        public static AtomicSafetyHandle GetSafetyHandle<T>(ref DeferredNativeArray<T> deferredNativeArray)
-            where T : struct
-        {
-            return deferredNativeArray.m_Safety;
-        }
-
-        /// <summary>
-        /// Gets the pointer to the safety handle for a <see cref="DeferredNativeArray{T}"/>
-        /// </summary>
-        /// <param name="deferredNativeArray">The instance to get the safety handle pointer from.</param>
-        /// <typeparam name="T">The type of <see cref="DeferredNativeArray{T}"/></typeparam>
-        /// <returns>The pointer to the <see cref="AtomicSafetyHandle"/> instance</returns>
-        public static void* GetSafetyHandlePointer<T>(ref DeferredNativeArray<T> deferredNativeArray)
-            where T : struct
-        {
-            return UnsafeUtility.AddressOf(ref deferredNativeArray.m_Safety);
         }
     }
 }
